Restructure: Auth0 package
- Introduce a layer (User) to group all logic related to
  "User Management" in the Auth0 package.
- Update the links in the documentation.
