+++ app/Kwality.UVault.QA/Management/User/Mappers/Auth0.User.Create.Operation.Mapper.Tests.cs
@@ -33,7 +33,7 @@ using FluentAssertions;
 using JetBrains.Annotations;
 
 using Kwality.UVault.QA.Internal.Xunit.Traits;
-using Kwality.UVault.Auth0.Operations.Mappers;
+using Kwality.UVault.Auth0.Users.Operations.Mappers;
 using Kwality.UVault.Users.Exceptions;
 using Kwality.UVault.Users.Operations.Mappers.Abstractions;
